requirements.txt
django>=5.1.1
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
